Adding support for mid-quantum preemption
The simulator now expects the following methods in a scheduler:
- job_quantum_expired
- job_preempted
- needs_resched

The first method is a logical replacement for the original `job_preempted`
and is called if a CPU burst runs until a time quantum expires (a MLFQ
implementation may then choose to lower the priority of the job).

The last method is now called by the simulator after a new job becomes
ready to check if the simulator believes a more deserving job exists. If
`needs_resched` returns true, then the simulator will arrange to preempt
the current job, after which it will invoke `job_preempted`.